+++ packages/terra-framework-docs/package.json
@@ -49,6 +49,7 @@
     "terra-form-input": "^4.4.0",
     "terra-form-select": "^6.8.0",
     "terra-grid": "^6.0.0",
+    "terra-hookshot": "^5.38.0",
     "terra-modal-manager": "^6.58.0",
     "terra-section-header": "^2.0.0",
     "terra-slide-panel-manager": "^5.61.0",
