+++ packages/terra-framework-docs/src/terra-dev-site/doc/hookshot/Hookshot.1.doc.mdx
@@ -2,9 +2,9 @@ import { Badge } from 'terra-hookshot/package.json?dev-site-package';
 
 import HookshotSrc from 'terra-hookshot/src/Hookshot?dev-site-props-table';
 import HookshotContentSrc from 'terra-hookshot/src/HookshotContent?dev-site-props-table';
-import HookshotExample from '../example/HookshotExample?dev-site-example';
-import BoundedHookshotExample from '../example/BoundedHookshotExample?dev-site-example';
-import CoordsHookshotExample from '../example/CoordsHookshotExample?dev-site-example';
+import HookshotExample from './example/HookshotExample?dev-site-example';
+import BoundedHookshotExample from './example/BoundedHookshotExample?dev-site-example';
+import CoordsHookshotExample from './example/CoordsHookshotExample?dev-site-example';
 
 <Badge />
