+++ packages/terra-framework-docs/src/terra-dev-site/doc/hookshot/example/BoundedHookshotExample.jsx
