+++ packages/terra-framework-docs/src/terra-dev-site/test/hookshot/HookshotCoordinates.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Hookshot from '../../../Hookshot';
-import HookshotContent from '../common/HookshotContentTestTemplate';
-import styles from '../common/HookshotTestDocCommon.module.scss';
+import Hookshot from 'terra-hookshot';
+import HookshotContent from './common/HookshotContentTestTemplate';
+import styles from './common/HookshotTestDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
