+++ packages/terra-hookshot/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs to terra-framework-docs
+
 ## 5.38.0 - (September 20, 2021)
 
 * Changed
